requirements.txt
alpaca-py
backtrader
google-cloud-pubsub